Fix DCHECK because the runloop did not have a task runner.
Due to the destruction order, the task execution environment in
App::Run() gets destroyed before the App object. Therefore, the run loop
is missing a task runner and dchecks.

This CL fixes flushing of prefs for UpdateApps just like they it was
fixed for InstallApp. It is not needed to wait for prefs to flush
under a runloop.

Bug: 1053729
Change-Id: I0ed8d3e97366c052277716e85cbaa74cad81236d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103320
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750374}