back gesture nudge: Better handle nudge not being initially available
Back gesture nudge would check whether the back nudge can be shown on
session change, or tablet mode change, and if that wasn't the case, it
would stop monitoring window changes. This way, if back nudge was not
shown initially, it would not get shown again unless tablet mode or
session state change, even if sufficient time to make the back nudge
available passed.

This cl updates the logic to schedule a delayed task to reevaluate
whether window monitoring should start if the nudge is not shown
initially. The task delay is provided by the
contextual_tooltip::ShouldShowNudge, and it indicates when the nudge is
expected to be showable again.

BUG=1061556

Change-Id: I4b961d1b6272903cb379086e9557886bd6239e5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103509
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750393}