Add -Wno-pointer-to-int-cast to cflags_c
The warning was changed to be emitted under -Wpointer-to-int-cast in
https://reviews.llvm.org/rG51d7f645448b49c7d3b029a60532bac019da1348.

Since -Wno-microsoft-cast is on right now with the pinned Clang, I think
it makes sense to not add this under llvm_force_head_revision. We can
remove -Wno-microsoft-cast after the pinned Clang contains the LLVM
change above.

Bug: 1054220
Change-Id: I2b989cbe4936fa237e7cfb2f725d14873d5abaf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104127
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750416}