[Web Payment] Remove minimal UI render test.
Before this patch, MicrotransactionRenderTest.java resided in the
repository, but was not included in the build, because it was removed in
https://crrev.com/c/1828291 due to build bot failure and then was
accidentally added back in https://crrev.com/c/1826875 without adding it
back to the build.

This patch removes MicrotransactionRenderTest.java file.

After this patch, the repository does not contain the unused file
MicrotransactionRenderTest.java.

Bug: 1005076
Change-Id: Ie143f174ddecacd53449b423bae82b570370f731
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093165
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750437}