Demo Mode: Record Demo Mode setup metrics
We want to keep records of how long operators need to wait for demo mode
to set up. Currently, there exists a metric of how long is spent on the
demo-setup page (OOBE.StepCompletionTime.Demo-setup) which has not yet
been added to the histograms.xml file. However, this metric is only
useful for durations up to 3 minutes, so we require a metric calculated
separately.

Furthermore, we want to know:
- How long is end-to-end setup
- How many times do operators need to retry during the setup process

Bug: b:148753889, 1061305
Change-Id: I9d8d656ca85863c3975419ab306c097b0bd00d4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094065
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Patti <patricialor@chromium.org>
Commit-Queue: Joseph Kim <josephkimsh@google.com>
Cr-Commit-Position: refs/heads/master@{#750452}