Record UMA when user changes shelf navigation buttons in tablet mode
Adds histograms that record value set by user for accessibility setting
to show shelf navigation buttons in tablet mode.
The action is recorded separately for OS a11y settings and OOBE.
The value set by user is recorded with a 10 second delay, to avoid
overreporting if the user toggles the setting few times in short
succession (when the value changes, the previous requests to record the
metric are canceled).

BUG=1055688

Change-Id: Ie8a8104e4feda6844632b3a249242163949098d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2099335
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750455}