viz: Add flag to enable splitting quads for occlusion
This patch adds a flag to enable/disable splitting quads when partially
occluded to reduce compositing overdraw.
This flag currently disables the optimization, so that we can consider
merging this patch back to 81 along with the optimization (then the
optimization will need to be explicitly enabled via the flag in 81).


Bug: 1061708
Test: Power consumption of meets on a chromebook with the flag on and off
Change-Id: I4417076a21c7074244e7bb960b1e1cf477fb721a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103721
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750460}