[mathml] Implement the CSS property math-style
Implement the CSS property math-style as specified in [1]. The property
is still a proposal [2] so it is toggled using the runtime flag CSSMathStyleEnabled.

It a CSS property rather than an extra field of ComputedStyle because it will be used (amongst other things) in the MathML UA sheet to implement
complex inheritance of scriptlevel/displaystyle and in the resolution of font-size (in particular they should have high priority).

Additionally, we believe it makes sense to expose these properties for
users (other than just MathML attributes) so that people can understand
the magic behind the scriptlevel and displaystyle changes. However, we
probably don't want to expose these properties yet, until details
are agreed with the CSSWG.

[1] https://mathml-refresh.github.io/mathml-core/#the-math-style-property
[2] https://mathml-refresh.github.io/mathml-css-proposals/math-script-level-and-math-style-explainer.html

Bug: 6606
Change-Id: I0d854c0829f5b7453901651a97efd5a7e5cea39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2100787
Commit-Queue: Rob Buis <rbuis@igalia.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Frédéric Wang <fwang@igalia.com>
Cr-Commit-Position: refs/heads/master@{#750505}