[CV] Treat extension with damaged hashes as corrupted
Before this CL if extension had any problem with computed_hashes.json
file, content verification will fail with MISSING_ALL_HASHES failure.
Previously it was not an issue, since:
 * With mode equals to ENFORCE_SCRICT, as in
ChromeContentVerifierDelegate, this failure reason makes no difference.
We still consider extension as corrupted.
 * For extensions from Chrome Web Store this is unusual situation: we
can recompute hashes and verify them using hashes signed by Chrome Web
Store.

But after work in https://crbug.com/958794 there is a strong difference
between cases "no computed_hashes.json file at all" and "there is the
file, but it is damaged and unreadable itself". In the first one we
should ignore the failure. In the future we will reinstall such an
extensions, but without disabling them before. But in the second case
damaged file with hashes may mean that other files are damaged too, so
we have to disable extension and reinstall it immediately.

This commit reflects mentioned difference, adding a new type of failure
and passing it from ComputedHashes::CreateFromFile to the VerifyFailed
callback in content verifier delegate.

Bug: 958794
Change-Id: Ibb2ef46f00b97e09aec593c44fceed93ad73e81f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062276
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750522}