[iOS] Fix crash on signout from managed account.
When the flag kClearSyncedData is enabled, the current code attempts to
add the new dialog option to the managed account alert. We only expect
the change to affect non-managed accounts, and so the app crashes.

Adds regression test for the case where a managed user signs out from
Chrome.

Bug: 1058662
Change-Id: Iaad9a13655a45653aa2e09e1428c03a37e412325
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101008
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750537}