Fix combining characters for c with acute accent
The table had ´c map to ç instead of ć, imitating GTK from circa 2008.
This was overlooked in the conversion, and left no way for users to
type ć (used in Polish for example).

I speculate that GTK was imitating the Windows ‘US International’ layout,
which has that mapping. We have a separate section of the combining table
for the Windows-like ‘US International PC’ that does map ´c map to ç,
but this should not still happen for other layouts.

Bug: 1061753
Change-Id: I113b3d6c3fda6b93f7f27296b1a30dae6aa0b668
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2102776
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Kevin Schoedel <kpschoedel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750545}