Avoid recomputing shouldCloseTab in InterceptNavigationDelegateImpl.java
This is just a readability improvement, no functional changes. It was
confusing that shouldCloseTab appeared to only be used in some cases
like actions that result in OVERRIDE_WITH_ASYNC_ACTION, when we were
actually just recomputing shouldCloseTab under the assumption that it
wouldn't change.

Change-Id: I52b16cef1483c328bc595bc329f4a9098b3fc5ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101311
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750575}