[sqlite] Switch to external SQLite repository.
This change shifts from building the SQLite source in Chromium
to building the SQLite source which now resides in its own
repository.

Motivation: now that SQLite has an official Git mirror we can
switch from manually maintaining a set of patches to having
a Chromium specific set of branches (master and releases).
This simplifies the maintenance task and this repository.

Bug: 945204
Change-Id: I83febc92d8968a75be69aeb18b352561b8817f0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013741
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/master@{#750587}