Re-enable TabletmodeTransitionTest
This has been disabled due to flakiness. The flakiness was
observed on asan bot and win7 bot. Now this test is only
built on non-sanitizer builds for chromeos (see r749747),
so it's safe to re-enable it.

As I tried running, it still fails sometimes on debug builds
on waiting logic, so I modified it slightly.

Bug: 1057864, 1057868
Test: interactive_ui_tests --gtest_repeat=20
Change-Id: I9e8112ffbd6f30f042ffd03f6b29bc7e1dbe18b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103521
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750605}