Add Visibility changed logic to IsolatedPrerenderTabHelper
Adds the following checks:
* Prefetching cannot start if the tab is not visible.
* When a tab becomes visible, prefetching is restarted.

Note that a tab becoming hidden does not stop the current prefetch
request. Instead, the current prefetch is allowed to finish in the
hidden tab but no more will begin. This is partly done to avoid
additional code complexity in storing the current prefetch url and
partly because a user is likely to switch between a SRP tab and its
results, so prefetching those links is beneficial in the end.

Bug: 1023486
Change-Id: I03cff0f77721f450ac97b4aec39ad87290cbc886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105403
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750609}