Separate out TabState fields to separate class and change serialization method to protobufs.
Currently in TabState fields are persisted by storing in a file one after
another. This approach does not scale well - fields are not well organized
and this approach makes it difficult to deprecate fields.

PersistedTabData enables persisted fields to be better organized and
protobufs offer better support for adding and deprecating fields.

PersistedTabData also enables different storage methods - in an
upcoming CL a PersistedTabData which uses a Level DB backend will
be used for data for building new features but is not pertinent
to the running of the app.

Bug: 1046367
Change-Id: Id73074303641babb24e2f6109254c53d0d763e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025317
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750611}