Animate login shelf gesture contextual nudge
Adds an animation to shelf drag handle and gesture contextual nudge for
login shelf gesture controller. The login shelf gesture handler is
active during the last step of OOBE, where the user can swipe up from
the shelf to start the session.

This CL adds an animation that moves the drag handle and the contextual
nudge up and down, to bring more attention to it.

Also, updates the drag handle size and color based on the updated
specification.

BUG=1061627

Change-Id: Iab380a84e8ff769ae16f82f5591fe9c8befaa980
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103655
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750616}