Enable the subresource filter for frames with a doc.write-aborted load
Subframes which have their initial load aborted (e.g. due to a
Document.write() call) currently never have the subresource filter
activated. We change the behavior so that a failed provisional load
triggers filter creation if a filter has not been already set. As no
navigation has occurred, we use the parent's activation in this case.

In line with this change, we rename a number of variables and
functions. While this change will affect some SubresourceFilter.*
histograms, we don't version them as we expect the number of affected
frames to be small.

Bug: 1052362
Change-Id: I6852fcf1ae72da92b2fe666b3c77c216209d8a87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2064474
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750649}