desktop-pwas: Create web-app launcher executable
This change adds chrome_pwa_launcher.exe, a small binary used to
launch Progressive Web Apps (PWAs). It will enable a PWA to register
as a handler for a file type by associating the file type with a
hardlink to or copy of (on file systems where hardlinks are
unsupported) chrome_pwa_launcher, created exclusively for that PWA.

Much like chrome_proxy, chrome_pwa_launcher runs Chrome with the
command-line arguments it is given. Unlike chrome_proxy,
chrome_pwa_launcher has an icon (a generic Chrome Web Store icon), to
be displayed for PWAs on Windows 7, where setting a custom icon is not
possible. chrome_pwa_launcher has no manifest, enabling its display
name to be customized.

Because chrome_pwa_launcher is a standalone executable run from the
User Data directory, it includes functionality to:
1. Find the appropriate chrome.exe to launch, done by reading the Last
   Browser file in the User Data directory;
2. Pass its version to Chrome, so that Chrome can update all PWA
   launchers if needed; and
3. Log its results to the registry, a lightweight approach to error
   detection that avoids the overhead of full Crashpad integration.

Each PWA's executable path will be recorded in the registry as:
   "<launcher hardlink or copy path>" "--app-id=<PWA app ID>" -- "%1"

Running chrome_pwa_launcher this way effectively launches Chrome with
the PWA's app ID as an argument. Using a separate, differently named
copy/hard link for each PWA is needed to prevent Windows from treating
PWAs in the registry as one (as it does if their paths point to the
same executable).

Bug: 960245
Change-Id: Ieef3d37c92c08e3638191580dc004f912d1c7fb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815220
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/master@{#750656}