Sketch makeCredential request handling
This changes the native caBLEv2 authenticator code to notify the
Java-side code of a processed makeCredential CTAP request. The Java code
then calls the privileged FIDO2 APIs to create a FIDO credential,
tells the native side to marshal it into a response message, which then
gets sent as a fidoStatus notification by the Java-side BLEHandler.

This is only a sketch because no request or response parameters are
actually forwarded over JNI yet: The request sent to the FIDO2 API and
the response passed back to the client are canned fakes.

Bug: 1002262
Change-Id: I9ed4c41e2fb6d5ce3f7c5655d9b687e3d269b504
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103556
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750662}