Improve readability of color space setting.
Currently, the calls to set the color spaces with and without
alpha are separate calls. It isn't terribly clear that the two
calls are identical except for the alpha setting.

This CL changes it to one call inside a loop over {true, false}.
A more simple godbolt.org test showed that on -O2 the compiler was
smart about optimizing it out to be the same thing. However, even
if the generated code is different, it should be insignificant.

Bug: 955158
Change-Id: Icb32001c1f6651ef643e7832539dafd4b69bf9d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2104991
Commit-Queue: Chris Blume <cblume@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750682}