[Offline indicator v2] Fix showing/hiding on tab switcher
When animating on a tab with web contents, we run Java the animations
and let the compositor animate the browser controls based on the browser
controls height changes. For now, we don't want to animate while on the
tab switcher. Instead, we will be skipping the animation and
repositioning the toolbar and the rest of the content immediately.

This CL changes the way we reposition the toolbar's Java view. Instead
of waiting for an offset from the ChromeFullscreenManager, we will
listen for height changes directly if we're not on a tab.

Also, the tab switcher will be listening to the browser controls height
changes to update its layout.

Bug: 1059438
Change-Id: Ie7d1dbc0591288d85466ba26d030378dc55829bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2092552
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#750684}